modularizing/refactoring backend routing
