+++ src/routers/user.js
@@ -1,0 +1,81 @@
+const express = require('express')
+const User = require('../models/user')
+const router = new express.Router()
+
+// Creates new user - JSON POST: name, email, password
+// localhost:3000/users
+router.post('/users', async (req, res) => {
+    const user = new User(req.body)
+    try {
+        await user.save()
+        res.status(201).send(user)
+    } catch (e) {
+        res.status(400).send(e)
+    }
+})
+
+// GET request to /users. calls all users
+router.get('/users', async (req, res) => {
+    try {
+        const users = await User.find({})
+        res.send(users)
+    } catch (e) {
+        res.status(500).send()
+    }
+})
+
+// fetch individual user by DB-ID
+// GET localhost:3000/users/{id}
+router.get('/users/:id', async (req, res) => {
+    const _id = req.params.id
+
+    try {
+        const user = await User.findById(_id)
+        
+        if (!user) {
+            return res.status(404).send()
+        }
+        res.send(user)
+
+    } catch (e) {
+        res.status(500).send()
+    }
+})
+
+// this updates users files
+router.patch('/users/:id', async (req, res) => {
+    const updates = Object.keys(req.body)
+    const allowedUpdates = ['name', 'email', 'password', 'age']
+    const isValidOperation = updates.every((update) => allowedUpdates.includes(update))
+
+    if (!isValidOperation) {
+        return res.status(400).send({ error: 'invalid updates' })
+    }
+
+    try {
+        const user = await User.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true })
+        if (!user) {
+            return res.status(404).send()
+        }
+        res.send(user)
+    } catch (e) {
+        res.status(400).send(e)
+    }
+})
+
+// user delete call to their ID. No parameters needed. 
+router.delete('/users/:id', async (req, res) => {
+    try {
+        const user = await User.findByIdAndDelete(req.params.id)
+
+        if (!user) {
+            return res.status(404).send()
+        }
+
+        res.send(user)
+    } catch (e) {
+        res.status(500).send()
+    }
+})
+
+module.exports = router
